### 2024-08-26 #7147 @editorialbot
[PRE REVIEW]: GWPopulation: Hardware agnostic population inference for compact binaries and beyond
**Submitting author:** @ColmTalbot (<a href="http://orcid.org/0000-0003-2053-5582">Colm Talbot</a>)
**Repository:** https://github.com/ColmTalbot/gwpopulation
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.1.2
