### 2024-10-22 d9ba549
removed if statement from gaussianchieffchip model
