### 2024-10-22 4e759b6
edits to function
