### 2024-10-30 0c26112
Make GaussianChiEffChiP jittable (#110)
* removed if statement from gaussianchieffchip model

* edits to function

* removed file
