### 2025-02-03 9a86bce
ENH: make truncnorm more robust
I had reports of bad normalization for truncnorm when the mean was many sigma away from the support. The new method more closely aligns with scipy.
