### 2025-02-05 #7753 @plaplant
@bruce-edelman @sterinaldi thanks for agreeing to review this submission! When you get a chance, please respond to this GitHub issue with:
This will generate your reviewer checklist of items to work through. As you notice issues, please feel free to point them out here in this thread. You can also make issues on [the upstream repository](https://github.com/ColmTalbot/gwpopulation) for more detailed/involved suggestions.

Let me know if you have any questions!
