### 2025-02-20 6e90889
Change conda provider in CI
