+++ .github/workflows/python-package.yml
@@ -33,11 +33,12 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
     - name: Setup conda
-      uses: s-weigand/setup-conda@v1
+      uses: conda-incubator/setup-miniconda@v3
       with:
-        update-conda: true
         python-version: ${{ matrix.python-version }}
-        conda-channels: anaconda, conda-forge
+        channels: conda-forge,defaults
+        auto-update-conda: true
+        mamba-version: "*"
     - name: Install dependencies
       run: |
         conda install pip setuptools
