+++ .github/workflows/python-package.yml
@@ -41,9 +41,9 @@ jobs:
         mamba-version: "*"
     - name: Install dependencies
       run: |
-        conda install pip setuptools
-        conda install --file requirements.txt
-        conda install --file test_requirements.txt
+        mamba install -n test pip setuptools
+        mamba install -n test --file requirements.txt
+        mamba install -n test --file test_requirements.txt
         conda run -n test python -m pre-commit install
     - name: Install gwpopulation
       run: |
