### 2025-02-20 62f9ab3
CI: fix pre-commit executable
