### 2025-02-20 f6df3e0
CI: update conda in docs job
