### 2025-02-20 e2ce674
CI: try activating conda
