### 2025-02-20 18cf076
CI: set python version for docs
