### 2025-02-20 99ec95a
CI: add pages requirements for docs
