### 2025-02-20 03c8171
CI: try fix sphinx command
