+++ .github/workflows/pages.yml
@@ -53,8 +53,8 @@ jobs:
         run: |
           cp examples/*.ipynb docs/examples
           cd docs
-          make clean
-          make html
+          make SPHINXBUILD="mamba run python -m sphinx" clean
+          make SPHINXBUILD="mamba run python -m sphinx" html
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
         with:
