### 2025-02-20 e03fb5a
CI: another attempt at reading docs
