### 2025-02-20 faa7c4a
CI: install gwpopulation for docs
