### 2025-02-20 c40382c
CI: fix CI failures (#114)
* Pin scipy version in test for bilby

* Change conda provider in CI

* Attempt to use correct conda env

* Install into correct env

* CI: fix pre-commit executable

* CI: update conda in docs job

* CI: try activating conda

* CI: set python version for docs

* CI: add pages requirements for docs

* CI: try fix sphinx command

* CI: another attempt at reading docs

* CI: install gwpopulation for docs
