+++ .github/workflows/python-package.yml
@@ -33,31 +33,32 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
     - name: Setup conda
-      uses: s-weigand/setup-conda@v1
+      uses: conda-incubator/setup-miniconda@v3
       with:
-        update-conda: true
         python-version: ${{ matrix.python-version }}
-        conda-channels: anaconda, conda-forge
+        channels: conda-forge,defaults
+        auto-update-conda: true
+        mamba-version: "*"
     - name: Install dependencies
       run: |
-        conda install pip setuptools
-        conda install --file requirements.txt
-        conda install --file test_requirements.txt
-        pre-commit install
+        mamba install -n test pip setuptools
+        mamba install -n test --file requirements.txt
+        mamba install -n test --file test_requirements.txt
+        conda run -n test pre-commit install
     - name: Install gwpopulation
       run: |
-        pip install .
+        conda run -n test python -m pip install .
     - name: List installed
       run: |
         conda list
     - name: Run pre-commit checks
       run: |
-        pre-commit run --all-files --verbose --show-diff-on-failure
-        jupyter nbconvert --clear-output --inplace examples/*.ipynb
+        conda run -n test pre-commit run --all-files --verbose --show-diff-on-failure
+        conda run -n test jupyter nbconvert --clear-output --inplace examples/*.ipynb
         git reset --hard
     - name: Test with pytest
       run: |
-        pytest --cov gwpopulation -ra --color yes --cov-report=xml --junitxml=pytest.xml
+        conda run -n test pytest --cov gwpopulation -ra --color yes --cov-report=xml --junitxml=pytest.xml
     - name: Publish coverage to Codecov
       uses: codecov/codecov-action@v4
       with:
@@ -65,7 +66,7 @@ jobs:
         files: coverage.xml
         flags: python${{ matrix.python-version }}
     - name: Coverage report
-      run: python -m coverage report --show-missing
+      run: conda run -n test python -m coverage report --show-missing
     - name: Upload test results
       if: always()
       uses: actions/upload-artifact@v4
