### 2025-02-24 27b2cb3
added mu_spin
