### 2025-02-24 6962a61
BUG: fix normalization for regular truncated normal
