### 2025-02-24 6582224
ENH: make truncnorm more robust (#113)
* ENH: make truncnorm more robust

I had reports of bad normalization for truncnorm when the mean was many sigma away from the support. The new method more closely aligns with scipy.

* FMT: black fixes

* BUG: fix normalization for regular truncated normal
