### 2025-03-06 4fe70e4
BUG: make select default scalar
cupy enforces that the default argument to `select` is a scalar. This brings us in line with that.
