### 2025-03-06 c9988c7
BUG: add missing branch to truncnorm
