### 2025-03-06 ed9bed9
BUG: make select default scalar (#118)
* BUG: make select default scalar

cupy enforces that the default argument to `select` is a scalar. This brings us in line with that.

* BUG: add missing branch to truncnorm
