### 2025-03-06 27a6bc9
Add flexibility to gaussian spin tilt (#115)
* added mu_spin

* Fixed tests

* changed unit test priors

* fixed test
