### 2025-03-06 73b1ba0
BUG: fix bugs in truncnorm for cupy (#119)
