+++ docs/guide/index.rst
@@ -7,6 +7,7 @@ User Guide
 .. toctree::
     :maxdepth: 2
 
+    installation
     backend
     cosmology
     experimental
