### 2025-03-19 #7753 @sterinaldi
Hi @ColmTalbot – I have some minor comments/questions regarding your submission.
1. On the author list: it looks like that several co-authors had contributed minimally to the repository (one or two minor commits). Did they contribute to the project in other ways, e.g. with derivations or ideas? I am not sure what is JOSS's policy in this case, maybe @plaplant can clarify?
2. There is a discrepancy between the Python version required by GWPopulation (>3.9) and the one required by its dependencies (>3.10 or 3.11), leading to runtime errors if the code is downloaded from the repository and installed. I opened an issue a few weeks ago.

Concerning the paper:
1. Some readers may be unfamiliar with the models (PL+Peak/+Spline etc) and likelihood implemented in GWPopulation. I would suggest that you add a couple of references in the Summary section;
2. I personally find the _Statement of Need_ section pretty involuted and difficult to read. I would suggest a rephrasing of such section to state more plainly and linearly the functionality of GWPopulation. Being this a merely stylistic comment, however, I do not consider it a showstopper for the review, unless @bruce-edelman shares the same opinion.

Once these comments have been addressed, I am happy to give green light to the submission
