### 2025-03-23 #7753 @plaplant
@sterinaldi thanks very much for completing your initial review! I appreciate the feedback to the authors and making issues on the upstream repo.

Regarding the question about authorship: JOSS's stated policy is available [here](https://joss.readthedocs.io/en/latest/submitting.html#authorship). The short version is that we do allow for non-code contributions to merit authorship, such as active project direction. We leave it up to the authors to decide who to include as a co-author. While we do not require an official statement about the contribution from individual authors, perhaps in this case @ColmTalbot can briefly do so in a response to this thread (following a format/categorization such as [the Contributor Role Taxonomy (CRediT)](https://credit.niso.org/) or some other comparable scheme).

@bruce-edelman thanks again for agreeing to review! It has been about 6 weeks since the review process began, so please submit your initial review as soon as possible. In this thread, please respond with
Then, you can work your way through the individual items and note any necessary changes (either in this thread or as issues on [the upstream repository](https://github.com/ColmTalbot/gwpopulation). Let me know if you have any questions!
