### 2025-04-02 #7753 @bruce-edelman
Hi @ColmTalbot, I only have very minor comments (beyond those laid out by @sterinaldi) for this submission.

- There currently does not appear to be installation instructions in the documentation or source repository README -- but it does look like there is an open PR to add them to docs
- There are citations for each of the other ancillary software packages mentioned (i.e. numpy, jax, cupy, emcee, etc) but no citation when mentioning numpyro in the GWInferno comparison snippet. NumPyro maintainers request people cite their NeurIPS paper (arxiv link: https://arxiv.org/abs/1912.11554)
- I think the Statement of Need section clearly lays out the problems GWPopulation is designed to solve but maybe slightly lacks on who the target audience is and the relation to other work. Although these points are sufficiently addressed in the Summary and State of the Field sections. Other than that I think this section reads fine (with very minor tweaks mentioned below) and does **not** in my personal opinion require a re-write for me to sign off on it.
- Optional suggestions on a few sentences that read a little awkwardly:
  - Line 48 reads weird -- suggest changing "need" -> "needed"
  - Use of "standard" twice in the sentence on lines 53/54 feels redundant

With the installation instructions added to the documentation and the numpyro citation added when mentioning it, I am more than happy to recommend the submission for publication.
