### 2025-04-09 32f6c8c
MAINT: bump official python requirement
