### 2025-04-09 2be4b82
TST: try testing with more recent python
