+++ .pre-commit-config.yaml
@@ -3,12 +3,12 @@ repos:
     rev: 22.3.0
     hooks:
       - id: black
-        language_version: python3
+        language_version: python3.11
 -   repo: https://github.com/ambv/black
     rev: 22.3.0
     hooks:
       - id: black-jupyter
-        language_version: python3
+        language_version: python3.11
 -   repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.2.0
     hooks:
