### 2025-04-09 7b02193
CI: move pre-commits to separate stage
