+++ .github/workflows/python-package.yml
@@ -18,6 +18,24 @@ on:
     - cron: "0 0 * * 0"
 
 jobs:
+  pre-commits:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v5
+      with:
+        python-version: "3.11"
+    - name: Install gwpopulation
+      run: python -m pip install .
+    - name: Run pre-commits
+      run: |
+        python -m pip install pre-commit jupyter nbconvert
+        pre-commit install
+        pre-commit run --all-files
+        jupyter nbconvert --clear-output --inplace examples/*.ipynb
+        git reset --hard
+
   build:
 
     runs-on: ubuntu-latest
@@ -44,18 +62,19 @@ jobs:
         mamba install -n test pip setuptools
         mamba install -n test --file requirements.txt
         mamba install -n test --file test_requirements.txt
-        conda run -n test pre-commit install
+        # conda run -n test pre-commit install
     - name: Install gwpopulation
       run: |
         conda run -n test python -m pip install .
     - name: List installed
       run: |
         conda list
-    - name: Run pre-commit checks
-      run: |
-        conda run -n test pre-commit run --all-files --verbose --show-diff-on-failure
-        conda run -n test jupyter nbconvert --clear-output --inplace examples/*.ipynb
-        git reset --hard
+    # - name: Run pre-commit checks
+    #   if: ${{ matrix.python-version }} == "3.11"
+    #   run: |
+    #     conda run -n test pre-commit run --all-files --verbose --show-diff-on-failure
+    #     conda run -n test jupyter nbconvert --clear-output --inplace examples/*.ipynb
+    #     git reset --hard
     - name: Test with pytest
       run: |
         conda run -n test pytest --cov gwpopulation -ra --color yes --cov-report=xml --junitxml=pytest.xml
