### 2025-04-09 dd190eb
TST: touch up github action
