### 2025-04-09 6a81fc7
MAINT: bump official python requirement (#122)
* MAINT: bump official python requirement

* TST: try testing with more recent python

* TST: remove scipy pin

* FMT: use specific python version for black

* CI: move pre-commits to separate stage

* TST: touch up github action

* [skip-ci] MAINT: add py313 to readme
