+++ setup.cfg
@@ -8,14 +8,13 @@ license_file = LICENSE.md
 long_description = file: README.md
 long_description_content_type = text/markdown
 classifiers =
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
     Programming Language :: Python :: 3.12
+    Programming Language :: Python :: 3.13
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
-python_requires = ">=3.8"
+python_requires = ">=3.10"
 
 [options]
 packages = find:
