+++ joss/paper.bib
@@ -297,3 +297,20 @@ archivePrefix = {arXiv},
     pages = "061102",
     year = "2022"
 }
+
+@ARTICLE{NumPyro,
+       author = {{Phan}, Du and {Pradhan}, Neeraj and {Jankowiak}, Martin},
+        title = "{Composable Effects for Flexible and Accelerated Probabilistic Programming in NumPyro}",
+      journal = {arXiv e-prints},
+     keywords = {Statistics - Machine Learning, Computer Science - Artificial Intelligence, Computer Science - Machine Learning, Computer Science - Programming Languages, I.2.5, G.3, I.2.5, G.3},
+         year = 2019,
+        month = dec,
+          eid = {arXiv:1912.11554},
+        pages = {arXiv:1912.11554},
+          doi = {10.48550/arXiv.1912.11554},
+archivePrefix = {arXiv},
+       eprint = {1912.11554},
+ primaryClass = {stat.ML},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv191211554P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
