### 2025-04-09 #7753 @ColmTalbot
Thanks everyone for the feedback!

> On the author list: it looks like that several co-authors had contributed minimally to the repository (one or two minor commits). Did they contribute to the project in other ways, e.g. with derivations or ideas? I am not sure what is JOSS's policy in this case, maybe @plaplant can clarify?

The commit volume for all of the authors listed may not be large, but all have contributed additionally to the conception/choices made to some aspects of the package, even if they did not provide final implementations. I am happy to provide more information if that is deemed necessary, but I think it would be a disservice for any of these people to not receive credit for the success of this package.

> There is a discrepancy between the Python version required by GWPopulation (>3.9) and the one required by its dependencies (>3.10 or 3.11), leading to runtime errors if the code is downloaded from the repository and installed. I opened an issue a few weeks ago.

Thanks for opening the issue, this was resolved in https://github.com/ColmTalbot/gwpopulation/pull/122.

> Some readers may be unfamiliar with the models (PL+Peak/+Spline etc) and likelihood implemented in GWPopulation. I would suggest that you add a couple of references in the Summary section;

I added a footnote referring readers to the appropriate paper for this jargon.

> There currently does not appear to be installation instructions in the documentation or source repository README -- but it does look like there is an open PR to add them to docs

Indeed, https://github.com/ColmTalbot/gwpopulation/pull/121 will address this.

> There are citations for each of the other ancillary software packages mentioned (i.e. numpy, jax, cupy, emcee, etc) but no citation when mentioning numpyro in the GWInferno comparison snippet. NumPyro maintainers request people cite their NeurIPS paper (arxiv link: https://arxiv.org/abs/1912.11554)

Whoops! I've added this citation.

> I personally find the Statement of Need section pretty involuted and difficult to read. I would suggest a rephrasing of such section to state more plainly and linearly the functionality of GWPopulation. Being this a merely stylistic comment, however, I do not consider it a showstopper for the review, unless @bruce-edelman shares the same opinion on the need of a rephrasing.

> I think the Statement of Need section clearly lays out the problems GWPopulation is designed to solve but maybe slightly lacks on who the target audience is and the relation to other work. Although these points are sufficiently addressed in the Summary and State of the Field sections. Other than that I think this section reads fine (with very minor tweaks mentioned below) and does not in my personal opinion require a re-write for me to sign off on it.
Optional suggestions on a few sentences that read a little awkwardly:
Line 48 reads weird -- suggest changing "need" -> "needed"
Use of "standard" twice in the sentence on lines 53/54 feels redundant

I've made a few changes to this section in https://github.com/ColmTalbot/gwpopulation/pull/98/commits/0354282128f136fd1076bd3f4d056538b8e2bf60, hopefully these are enough to address the concerns.

- I've addressed the two specific points Bruce made.
- I added an additional clause stating main the userbase for the package.
