+++ .github/workflows/joss-paper.yml
@@ -17,7 +17,7 @@ jobs:
           journal: joss
           paper-path: joss/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           path: joss/paper.pdf
