+++ docs/guide/installation.rst
@@ -5,13 +5,16 @@ Installation
    .. tab:: Conda
       .. code-block:: console
           $ conda install -c conda-forge gwpopulation
-      Supported python versions: 3.10-3.12.
+      Supported python versions: 3.10-3.13.
    .. tab:: Pip
       .. code-block:: console
           $ pip install gwpopulation
-      Supported python versions: 3.10-3.12.
+      Supported python versions: 3.10-3.13.
 
-This will install the 
+This will install the latest stable release of `gwpopulation` into your
+python package directory, including all required dependencies (example, `bilby`
+and `dynesty`) if they are not already available in your system. Methods from
+`gwpopulation` should then be accessible for import into your program file.
 
 `gwpopulation` source installation
 --------------------------------
@@ -26,7 +29,6 @@ main software.
 
    $ git clone git@github.com:ColmTalbot/gwpopulation.git
    $ cd gwpopulation/
-   $ pip install -r requirements.txt
    $ pip install .
 
 Once the installation is finished, you can check that the installation proceeded
