### 2025-04-21 f656759
Add Installation Instructions to the Docs (#121)
* Create installation.txt

* add installation page

* review installation instructions
