### 2025-04-22 #7753 @Suyog7130
> > On the author list: it looks like that several co-authors had contributed minimally to the repository (one or two minor commits). Did they contribute to the project in other ways, e.g. with derivations or ideas? I am not sure what is JOSS's policy in this case, maybe [@plaplant](https://github.com/plaplant) can clarify?

> > There currently does not appear to be installation instructions in the documentation or source repository README -- but it does look like there is an open PR to add them to docs
>
> Indeed, [ColmTalbot/gwpopulation#121](https://github.com/ColmTalbot/gwpopulation/pull/121) will address this.
>

Hii, dunno if this would be of any use, but if ya are okayy with adding people to the author list who have made minor contributions to the package, then @ColmTalbot, maybe my name can be added as well (I made a small contribution).
