### 2025-04-22 #7753 @ColmTalbot
> Hii, dunno if this would be of any use, but if ya are okayy with adding people to the author list who have made minor contributions to the package, then @ColmTalbot, maybe my name can be added as well (I made a small contribution).

Hi @Suyog7130 I appreciate your help with the documentation, but I don't think this would be appropriate at this stage given the contribution.
