### 2025-04-22 b8e59ea
DOC: fix installation docs
