+++ docs/conf.py
@@ -20,6 +20,7 @@ extensions = [
     "myst_parser",
     "sphinx_design",
     "sphinx_copybutton",
+    'sphinx_tabs.tabs',
 ]
 
 templates_path = ["_templates"]
