+++ docs/guide/installation.rst
@@ -1,6 +1,7 @@
 ============
 Installation
 ============
+
 .. tabs::
    .. tab:: Conda
       .. code-block:: console
@@ -11,27 +12,41 @@ Installation
           $ pip install gwpopulation
       Supported python versions: 3.10-3.13.
 
-This will install the latest stable release of `gwpopulation` into your
-python package directory, including all required dependencies (example, `bilby`
-and `dynesty`) if they are not already available in your system. Methods from
-`gwpopulation` should then be accessible for import into your program file.
+This will install the latest stable release of :code:`gwpopulation` into your
+python package directory, including all required dependencies (e.g.,, :code:`bilby`
+and :code:`dynesty`) if they are not already available in your system. Methods from
+:code:`gwpopulation` should then be accessible for import into your program file.
 
-`gwpopulation` source installation
---------------------------------
+Additional backends
+-------------------
 
-Since updates and bug corrections to `gwpopulation` are being constantly made,
-it might be a good idea to install the latest available version from the source.
+The main power of :code:`gwpopulation` is the ability to use non-:code:`numpy` backends
+for array operations.
+If installing using :code:`pip`, the optional requirements for specifiy backends can be installed
+by specifying, e.g.,
+
+.. code-block:: console
+
+   $ pip install gwpopulation[jax]
+
+:code:`gwpopulation` source installation
+----------------------------------------
+
+A development version of :code:`gwpopulation` can be installed from the source code
+in the usual way.
 Assuming a working python installation exists, this can be done by cloning the
-`gwpopulation` repository, installing the requirements and then installing the
-main software.
+:code:`gwpopulation` repository and installing using :code:`pip`.
 
 .. code-block:: console
 
    $ git clone git@github.com:ColmTalbot/gwpopulation.git
    $ cd gwpopulation/
-   $ pip install .
+   $ pip install -e .
+
+.. note::
+
+   Installing with `-e` will produce an `editable installation <https://setuptools.pypa.io/en/latest/userguide/development_mode.html>`_.
+   Additional backends can also be specified via editable installations using :code:`pip install -e .[jax]`.
 
 Once the installation is finished, you can check that the installation proceeded
 successfully by running one of the examples outlined on the Examples page.
-
-
