### 2025-04-22 2fc1a24
DOC: refactor docs page
