### 2025-04-22 5c27f70
DOC: typo fix in installation instructions
