### 2025-04-22 0a0b137
DOC: fix installation docs (#125)
* DOC: fix installation docs

* TYPO: precommit fixes

* DOC: fix whitespace errors

* DOC: refactor docs page

* DOC: typo fix in installation instructions
