+++ pages_requirements.txt
@@ -12,3 +12,4 @@ sphinx_rtd_theme
 sphinx-autoapi
 sphinx-copybutton
 sphinx-design
+sphinx-tabs
