### 2025-04-22 #7753 @ColmTalbot
> Double check authors and affiliations (including ORCIDs)

Done

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

I just created [v1.2.1](https://github.com/ColmTalbot/gwpopulation/releases/tag/v1.2.1)

> Archive the release on Zenodo/figshare/etc and post the DOI here.

[Zenodo link](https://zenodo.org/records/15264649) for the new release

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

The zenodo authors pulls automatically from the full repository contributor list which differs slightly from the listing on the paper.

> Make sure that the license listed for the archive is the same as the software license.

Done
