### 2025-05-15 #7753 @plaplant
@ColmTalbot thanks for the information! Apologies for the delay in responding, I was running into some formatting issues on the manuscript. I've made a [small PR](https://github.com/ColmTalbot/gwpopulation/pull/126) that will address some formatting and style issues in the manuscript. When that's been merged in, we can move forward with acceptance. Thanks for your patience!
