### 2025-05-15 7d7d9c5
Make some small wording changes; add full author list for publications (#126)
