+++ joss/paper.bib
@@ -1,8 +1,8 @@
 @article{GW150914,
-  title = {Observation of Gravitational Waves from a Binary Black Hole Merger},
-  author = {Abbott, B. P. and {et al.}},
+  title = "{Observation of Gravitational Waves from a Binary Black Hole Merger}",
+  author = {Abbott, B. P. and Abbott, R. and Abbott, T. D. and Abernathy, M. R. and Acernese, F. and Ackley, K. and Adams, C. and Adams, T. and Addesso, P. and Adhikari, R. X. and Adya, V. B. and Affeldt, C. and Agathos, M. and Agatsuma, K. and Aggarwal, N. and Aguiar, O. D. and Aiello, L. and Ain, A. and Ajith, P. and Allen, B. and Allocca, A. and Altin, P. A. and Anderson, S. B. and Anderson, W. G. and Arai, K. and Arain, M. A. and Araya, M. C. and Arceneaux, C. C. and Areeda, J. S. and Arnaud, N. and Arun, K. G. and Ascenzi, S. and Ashton, G. and Ast, M. and Aston, S. M. and Astone, P. and Aufmuth, P. and Aulbert, C. and Babak, S. and Bacon, P. and Bader, M. K. M. and Baker, P. T. and Baldaccini, F. and Ballardin, G. and Ballmer, S. W. and Barayoga, J. C. and Barclay, S. E. and Barish, B. C. and Barker, D. and Barone, F. and Barr, B. and Barsotti, L. and Barsuglia, M. and Barta, D. and Bartlett, J. and Barton, M. A. and Bartos, I. and Bassiri, R. and Basti, A. and Batch, J. C. and Baune, C. and Bavigadda, V. and Bazzan, M. and Behnke, B. and Bejger, M. and Belczynski, C. and Bell, A. S. and Bell, C. J. and Berger, B. K. and Bergman, J. and Bergmann, G. and Berry, C. P. L. and Bersanetti, D. and Bertolini, A. and Betzwieser, J. and Bhagwat, S. and Bhandare, R. and Bilenko, I. A. and Billingsley, G. and Birch, J. and Birney, I. A. and Birnholtz, O. and Biscans, S. and Bisht, A. and Bitossi, M. and Biwer, C. and Bizouard, M. A. and Blackburn, J. K. and Blair, C. D. and Blair, D. G. and Blair, R. M. and Bloemen, S. and Bock, O. and Bodiya, T. P. and Boer, M. and Bogaert, G. and Bogan, C. and Bohe, A. and Bojtos, P. and Bond, C. and Bondu, F. and Bonnand, R. and Boom, B. A. and Bork, R. and Boschi, V. and Bose, S. and Bouffanais, Y. and Bozzi, A. and Bradaschia, C. and Brady, P. R. and Braginsky, V. B. and Branchesi, M. and Brau, J. E. and Briant, T. and Brillet, A. and Brinkmann, M. and Brisson, V. and Brockill, P. and Brooks, A. F. and Brown, D. A. and Brown, D. D. and Brown, N. M. and Buchanan, C. C. and Buikema, A. and Bulik, T. and Bulten, H. J. and Buonanno, A. and Buskulic, D. and Buy, C. and Byer, R. L. and Cabero, M. and Cadonati, L. and Cagnoli, G. and Cahillane, C. and Bustillo, J. Calder\'on and Callister, T. and Calloni, E. and Camp, J. B. and Cannon, K. C. and Cao, J. and Capano, C. D. and Capocasa, E. and Carbognani, F. and Caride, S. and Diaz, J. Casanueva and Casentini, C. and Caudill, S. an
[…]
