### 2025-05-15 #7753 @plaplant
@ColmTalbot thanks for merging the paper PR. For publication in JOSS, [we require wherever possible that the archived version of the software have the same title as the manuscript and that the author list matches](https://joss.readthedocs.io/en/latest/editing.html#post-review). When you get a chance, can you update the name of the Zenodo page and the author list to match the submission? I understand that the author list of Zenodo is generated automatically, but it's also possible manually set the metadata and shouldn't require making a new release. Thanks!
