### 2025-05-29 #7753 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
