### 2025-05-29 ab61953
Initial editorial tweaks for JOSS paper (#127)
