### 2025-05-29 4d777a5
Initial editorial tweaks for JOSS paper (#127)
