+++ joss/paper.md
@@ -1,8 +1,8 @@
 ---
 title: 'GWPopulation: Hardware agnostic population inference for compact binaries and beyond'
 tags:
-  -  Python
-  -  astronomy
+  - Python
+  - astronomy
   - gravitational waves
   - hierarchical inference
 authors:
@@ -22,17 +22,17 @@ authors:
      orcid: 0000-0002-4534-0485
      affiliation: 7, 8
 affiliations:
-  -  name: Kavli Institute for Cosmological Physics, University of Chicago, USA
+  -  name: Kavli Institute for Cosmological Physics, University of Chicago, Chicago, IL 60637, USA
      index: 1
-  -  name: Department of Physics, University of Chicago, Chicago, USA
+  -  name: Department of Physics, University of Chicago, Chicago, IL 60637, USA
      index: 2
-  -  name: Laboratoire Lagrange, Université Côte d'Azur, Observatoire de la Côte d'Azur, CNRS,  Bd de l'Observatoire, 06300, France
+  -  name: Laboratoire Lagrange, Université Côte d'Azur, Observatoire de la Côte d'Azur, CNRS, Bd de l'Observatoire, 06300, France
      index: 3
-  -  name: Laboratoire Artemis, Université Côte d'Azur, Observatoire de la Côte d'Azur, CNRS,  Bd de l'Observatoire, 06300, France
+  -  name: Laboratoire Artemis, Université Côte d'Azur, Observatoire de la Côte d'Azur, CNRS, Bd de l'Observatoire, 06300, France
      index: 4
-  -  name: Department of Physics, California Institute of Technology, Pasadena, CA
+  -  name: Department of Physics, California Institute of Technology, Pasadena, CA 91125, USA
      index: 5
-  -  name: LIGO Laboratory, California Institute of Technology, Pasadena, CA
+  -  name: LIGO Laboratory, California Institute of Technology, Pasadena, CA 91125, USA
      index: 6
   -  name: School of Physics and Astronomy, Monash University, VIC 3800, Australia
      index: 7
@@ -46,46 +46,46 @@ bibliography: paper.bib
 # Summary
 
 Since the first direct detection of gravitational waves by the LIGO--Virgo collaboration in 2015 [@GW150914], the size of the gravitational-wave transient catalog has grown to nearly 100 events [R. Abbott et al., -@GWTC3], with the ongoing fourth observing run more than doubling the total number.
-Extracting astrophysical/cosmological information from these observations is a hierarchical Bayesian inference problem.
-`GWPopulation` is designed to provide simple-to-use, robust, and extensible tools for hierarchical inference in gravitational-wave astronomy/cosmology. It has been widely adopted for gravitational-wave astrnomy, including producing flagship results for the LIGO-Virgo-KAGRA collaborations [R. Abbott et al., -@GWTC3Pop; @GW230529] [^1].
+Extracting astrophysical or cosmological information from these observations is a hierarchical Bayesian inference problem.
+`GWPopulation` is designed to provide simple-to-use, robust, and extensible tools for hierarchical inference in gravitational-wave astronomy or cosmology. It has been widely adopted for gravitational-wave astronomy, including producing flagship results for the LIGO-Virgo-KAGRA collaborations [R. Abbott et al., -@GWTC3Pop; @GW230529][^1].
 While designed to work with observations of compact binary coalescences, `GWPopulation` may be available to a wider range of hierarchical Bayesian inference problems.
 
 [^1]: For a full listing of papers using `GWPopulation`, see the [citations for the previous publication](https://ui.adsabs.harvard.edu/abs/2019PhRvD.100d3030T/citations).
 
-Building on `Bilby` [@Bilby], `GWPopulation` can easily be used with a range of stochastic samplers through a standard interface.
[…]
