### 2025-05-29 #7753 @warrickball
Thanks once again to @bruce-edelman and @sterinaldi for reviewing and @plaplant for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.
